finalized dream2

+++ Profiles/Assets/Source/DreamSymbols/ImageDrift.cs
@@ -5,7 +5,7 @@ public class ImageDrift : MonoBehaviour {
 
 	//This script takes a UI image and shakes it around in 2D directions to make look cool
 
-	public int shakeAmount;
+	public float shakeAmount;
 	public float shakeSpeed;
 
 	void Start()
@@ -17,8 +17,8 @@ public class ImageDrift : MonoBehaviour {
 	{
 		while (true)
 		{
-			int xAmount = Random.Range(-shakeAmount, shakeAmount);
-			int yAmount = Random.Range(-shakeAmount, shakeAmount);
+			float xAmount = (float)Random.Range(-shakeAmount, shakeAmount);
+			float yAmount = (float)Random.Range(-shakeAmount, shakeAmount);
 			gameObject.transform.Translate(new Vector3(xAmount, yAmount, 0), gameObject.transform);
 			yield return new WaitForSeconds(shakeSpeed);
 			gameObject.transform.Translate(new Vector3(-xAmount, -yAmount, 0), gameObject.transform);
